rename shooter handoff to feeder
